chardev: Let FDChardev::max_size be unsigned
Suggested-by: Paolo Bonzini <pbonzini@redhat.com>

+++ chardev/char-pty.c
@@ -34,7 +34,7 @@
 typedef struct {
     Chardev parent;
     QIOChannel *ioc;
-    int read_bytes;
+    size_t read_bytes;
 
     /* Protected by the Chardev chr_write_lock.  */
     int connected;
